omnigraphdiff/requirements.txt
torch>=2.0.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm>=4.62.0
lifelines>=0.27.0
umap-learn>=0.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.8.0
h5py>=3.6.0
